idea merge  - working on fs merge  - util map merge  - initial bus merge

+++ modules/base/platform-impl/src/main/java/com/intellij/openapi/vfs/newvfs/persistent/VfsDependentEnum.java
@@ -16,51 +16,52 @@
 package com.intellij.openapi.vfs.newvfs.persistent;
 
 import com.intellij.openapi.util.io.FileUtil;
-import com.intellij.util.containers.ConcurrentList;
 import com.intellij.util.containers.ContainerUtil;
 import com.intellij.util.io.DataExternalizer;
 import com.intellij.util.io.DataInputOutputUtil;
 import com.intellij.util.io.KeyDescriptor;
 import gnu.trove.THashMap;
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 
 import java.io.*;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.ConcurrentMap;
 
 // Relatively small T <-> int mapping for elements that have int numbers stored in vfs, similar to PersistentEnumerator<T>,
 // unlike later numbers assigned to T are consequent and retained in memory / expected to be small.
 // Vfs invalidation will rebuild this mapping, also any exception with the mapping will cause rebuild of the vfs
 // stored data is VfsTimeStamp Version T*
-//
 public class VfsDependentEnum<T> {
   private static final String DEPENDENT_PERSISTENT_LIST_START_PREFIX = "vfs_enum_";
   private final File myFile;
   private final DataExternalizer<T> myKeyDescriptor;
-  private int myVersion;
+  private final int myVersion;
 
   // GuardedBy("myLock")
   private boolean myMarkedForInvalidation;
 
-  private final ConcurrentList<T> myInstances = ContainerUtil.createConcurrentList();
-  private final ConcurrentMap<T, Integer> myInstanceToId = ContainerUtil.newConcurrentMap();
+  private final List<T> myInstances = ContainerUtil.createConcurrentList();
+  private final Map<T, Integer> myInstanceToId = ContainerUtil.newConcurrentMap();
   private final Object myLock = new Object();
   private boolean myTriedToLoadFile;
 
-  public VfsDependentEnum(String fileName, KeyDescriptor<T> descriptor, int version) {
-    myFile = new File(FSRecords.basePath(), DEPENDENT_PERSISTENT_LIST_START_PREFIX + fileName  + FSRecords.VFS_FILES_EXTENSION);
+  public VfsDependentEnum(@Nonnull String fileName, @Nonnull KeyDescriptor<T> descriptor, int version) {
+    myFile = new File(FSRecords.basePath(), DEPENDENT_PERSISTENT_LIST_START_PREFIX + fileName + FSRecords.VFS_FILES_EXTENSION);
     myKeyDescriptor = descriptor;
     myVersion = version;
   }
 
+  @Nonnull
   static File getBaseFile() {
     return new File(FSRecords.basePath(), DEPENDENT_PERSISTENT_LIST_START_PREFIX);
   }
 
   public int getId(@Nonnull T s) throws IOException {
+    return getIdRaw(s, true);
+  }
+
+  int getIdRaw(@Nonnull T s, boolean vfsRebuildOnException) throws IOException {
     Integer integer = myInstanceToId.get(s);
     if (integer != null) return integer;
 
@@ -81,47 +82,48 @@ public class VfsDependentEnum<T> {
         return enumerated;
       }
       catch (IOException e) {
-        throw invalidate(e);
+        invalidate(e, vfsRebuildOnException);
+        throw e;
       }
     }
   }
 
   private void saveToFile(@Nonnull T instance) throws IOException {
     FileOutputStream fileOutputStream = new FileOutputStream(myFile, true);
-    DataOutputStream output = new DataOutputStream(new BufferedOutputStream(fileOutputStream));
 
-    try {
+    try (DataOutputStream output = new DataOutputStream(new BufferedOutputStream(fileOutputStream))) {
       if (myFile.length() == 0) {
         DataInputOutputUtil.writeTIME(output, FSRecords.getCreationTimestamp());
         DataInputOutputUtil.writeINT(output, myVersion);
       }
       myKeyDescriptor.save(output, instance);
-    } finally {
+    }
+    finally {
       try {
-        output.close();
         fileOutputStream.getFD().sync();
-      }  catch (IOException ignore) {}
+      }
+      catch (IOException ignored) {
+      }
     }
   }
 
   private boolean loadFromFile() throws IOException {
-    if (!myTriedToLoadFile && myInstances.size() == 0 && myFile.exists()) {
+    if (!myTriedToLoadFile && myInstances.isEmpty() && myFile.exists()) {
       myTriedToLoadFile = true;
-      DataInputStream input = new DataInputStream(new BufferedInputStream(new FileInputStream(myFile)));
-      long vfsVersion = DataInputOutputUtil.readTIME(input);
+      boolean deleteFile = false;
+      try (DataInputStream input = new DataInputStream(new BufferedInputStream(new FileInputStream(myFile)))) {
+        long vfsVersion = DataInputOutputUtil.readTIME(input);
 
-      if (vfsVersion != FSRecords.getCreationTimestamp()) {
-        // vfs was rebuilt, so the list will be rebuit
-        try { input.close(); } catch (IOException ignore) {}
-        FileUtil.deleteWithRenaming(myFile);
-        return false;
-      }
+        if (vfsVersion != FSRecords.getCreationTimestamp()) {
+          // vfs was rebuilt, so the list will be rebuilt
+          deleteFile = true;
+          return false;
+        }
 
-      List<T> elements = new ArrayList<T>();
-      Map<T, Integer> elementToIdMap = new THashMap<T, Integer>();
-      int savedVersion = DataInputOutputUtil.readINT(input);
-      try {
+        int savedVersion = DataInputOutputUtil.readINT(input);
         if (savedVersion == myVersion) {
+          List<T> elements = new ArrayList<>();
+          Map<T, Integer> elementToIdMap = new THashMap<>();
           while (input.available() > 0) {
             T instance = myKeyDescriptor.read(input);
             assert instance != null;
@@ -131,31 +133,31 @@ public class VfsDependentEnum<T> {
           myInstances.addAll(elements);
           myInstanceToId.putAll(elementToIdMap);
           return true;
-        } else {
+        }
+        else {
           // force vfs to rebuild
           throw new IOException("Version mismatch: current " + myVersion + ", previous:" + savedVersion + ", file:" + myFile);
         }
       }
       finally {
-        try { input.close(); } catch (IOException ignore) {}
+        if (deleteFile) {
+          FileUtil.deleteWithRenaming(myFile);
+        }
       }
     }
     return false;
   }
 
   // GuardedBy("myLock")
-  private @Nullable IOException invalidate(@Nullable Throwable e) {
+  private void invalidate(@Nonnull Throwable e, boolean vfsRebuildOnException) {
     if (!myMarkedForInvalidation) {
-      doInvalidation(e);
       myMarkedForInvalidation = true;
+      // exception will be rethrown in this call
+      FileUtil.deleteWithRenaming(myFile); // better alternatives ?
+      if (vfsRebuildOnException) {
+        FSRecords.requestVfsRebuild(e);
+      }
     }
-    if (e instanceof IOException) return (IOException)e;
-    return null;
-  }
-
-  protected void doInvalidation(Throwable e) {
-    FileUtil.deleteWithRenaming(myFile); // better alternatives ?
-    FSRecords.requestVfsRebuild(e);
   }
 
   private void register(@Nonnull T instance, int id) {
@@ -164,8 +166,8 @@ public class VfsDependentEnum<T> {
     myInstances.add(instance);
   }
 
-  public @Nonnull
-  T getById(int id) throws IOException {
+  @Nonnull
+  public T getById(int id) throws IOException {
     assert id > 0;
     --id;
     T instance;
@@ -189,10 +191,8 @@ public class VfsDependentEnum<T> {
         }
         assert false : "Reading nonexistent value:" + id + "," + myFile + ", loaded:" + loaded;
       }
-      catch (IOException e) {
-        throw invalidate(e);
-      } catch (AssertionError e) {
-        invalidate(e);
+      catch (IOException | AssertionError e) {
+        invalidate(e, true);
         throw e;
       }
     }
